+++ Integrations/VS-Code/CHANGELOG.md
@@ -19,3 +19,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Compare with Remote native diff viewer.
 - Class generation command support for instances.
 - Context menu item for pushing the currently active or selected script to the server.
+- Context menu item for pulling the currently active or selected script from the server.